move the DataCollector (object that analyze a register) from the CLI script into a separate module into gnrwsgisite_proxy, where the current register resides. The collector is available via site object.
Update CLI and prometheus webtool accordingly.
